add availability toast
