+++ automerge/packages/jupyterlab-rtc/package.json
@@ -54,7 +54,9 @@
     "@material-ui/core": "4.11.2",    
     "@material-ui/icons": "4.11.2",    
     "@material-ui/lab": "4.0.0-alpha.57",    
-    "automerge": "file:./../../externals/automerge"
+    "automerge": "file:./../../externals/automerge",
+    "notistack": "1.0.3",
+    "jupyterlab_toastify": "4.1.3"
   },
   "devDependencies": {
     "@jupyterlab/builder": "^3.0.0",
